+++ .esdoc.json
@@ -3,7 +3,6 @@
   "destination": "./esdoc",
   "test": {
 	"type": "mocha",
-	"includes": ["(spec|specs|Spec|test|Test)\\.(js|es6)$"],
 	"source": "./test"
   }
 }
