+++ package.json
@@ -12,9 +12,9 @@
   "scripts": {
     "start": "gulp watch",
     "build": "gulp default",
-    "test": "mocha --inline-diffs --compilers js:babel-register './test/**/*.specs.js'",
-    "testWatch": "mocha --watch --compilers js:babel-register --watch-extensions dot './test/**/*.specs.js'",
-    "coverage": "babel-node ./node_modules/.bin/isparta cover ./node_modules/.bin/_mocha -- './test/**/*.specs.js'",
+    "test": "mocha --inline-diffs --compilers js:babel-register './test/**/*.spec.js'",
+    "testWatch": "mocha --watch --compilers js:babel-register --watch-extensions dot './test/**/*.spec.js'",
+    "coverage": "babel-node ./node_modules/.bin/isparta cover ./node_modules/.bin/_mocha -- './test/**/*.spec.js'",
     "coveralljs": "npm run coverage && cat ./coverage/lcov.info | node_modules/coveralls/bin/coveralls.js && rm -rf coverage",
     "esdoc": "esdoc -c .esdoc.json"
   },
