+++ test/cfg/cfg-builder.integration.spec.js
