+++ test/cfg/cfg-builder.spec.js
