+++ test/cfg/chai-path-helpe.js
