+++ test/cfg/control-flow-graph.spec.js
