+++ test/cfg/dot.spec.js
