+++ test/cfg/edge.spec.js
