+++ test/cfg/node.spec.js
