+++ test/semantic-model/parent-initializer-visitor.spec.js
