+++ test/semantic-model/scope.spec.js
