+++ test/semantic-model/source-file.spec.js
