+++ test/semantic-model/symbol-extractor.spec.js
