+++ test/semantic-model/symbol-table.spec.js
