+++ test/semantic-model/symbol.spec.js
