+++ test/semantic-model/type/function-type.spec.js
