+++ test/semantic-model/type/parametrized-type.spec.js
