+++ test/semantic-model/type/record-type.spec.js
