+++ test/semantic-model/type/type.spec.js
