+++ test/type-inference/forward-analysis-type-inference.integration.spec.js
