+++ test/type-inference/forward-analysis-type-inference.spec.js
