+++ test/type-inference/hindley-milner.spec.js
