+++ test/type-inference/refinement-context.spec.js
