+++ test/type-inference/refinment-rules/assignment-expression-refinement-rule.spec.js
