+++ test/type-inference/refinment-rules/binary-expression-refinement-rule.spec.js
