+++ test/type-inference/refinment-rules/binary-operators.spec.js
