+++ test/type-inference/refinment-rules/block-statement-refinement-rule.spec.js
