+++ test/type-inference/refinment-rules/expression-statement-refinement-rule.spec.js
