+++ test/type-inference/refinment-rules/function-refinement-rule.spec.js
