+++ test/type-inference/refinment-rules/identifier-refinement-rule.spec.js
