+++ test/type-inference/refinment-rules/if-statement-refinement-rule.spec.js
