+++ test/type-inference/refinment-rules/literal-refinement-rule.spec.js
