+++ test/type-inference/refinment-rules/member-expression-refinement-rule.spec.js
