+++ test/type-inference/refinment-rules/object-expression-refinement-rule.spec.js
