+++ test/type-inference/refinment-rules/return-statement-refinement-rule.spec.js
