+++ test/type-inference/refinment-rules/unary-statement-refinement-rule.spec.js
