+++ test/type-inference/refinment-rules/variable-declaration-refinement-rule.spec.js
