+++ test/type-inference/type-environment.spec.js
