+++ test/type-inference/unification-rules/record-unification-rule.spec.js
